fix(auth): clear verified signed-out session and share isSupabaseAuthApiError
