+++ apps/practitioner/src/lib/auth-callback-fragment-helpers.spec.ts
@@ -1,7 +1,4 @@
-import {
-  isSupabaseAuthApiError,
-  parseImplicitHashParams,
-} from './auth-callback-fragment-helpers';
+import { parseImplicitHashParams } from './auth-callback-fragment-helpers';
 
 describe('parseImplicitHashParams', () => {
   it('parses access_token and refresh_token from a typical invite hash', () => {
@@ -20,20 +17,5 @@ describe('parseImplicitHashParams', () => {
       access_token: 'a',
       refresh_token: 'b',
     });
-  });
-});
-
-describe('isSupabaseAuthApiError', () => {
-  it('returns true for Supabase Auth API errors', () => {
-    expect(
-      isSupabaseAuthApiError({
-        __isAuthError: true,
-        message: 'Network error',
-      }),
-    ).toBe(true);
-  });
-
-  it('returns false for ordinary errors', () => {
-    expect(isSupabaseAuthApiError(new Error('nope'))).toBe(false);
   });
 });
