+++ apps/practitioner/src/lib/auth-provider.spec.tsx
@@ -292,6 +292,52 @@ describe('AuthProvider', () => {
     expect(signOutMock).not.toHaveBeenCalled();
   });
 
+  it('clears session when verify returns no session without error', async () => {
+    const session = makeSession('practitioner-1');
+
+    getVerifiedAuthSessionMock
+      .mockResolvedValueOnce({
+        data: { user: null, session: null },
+        error: null,
+      })
+      .mockResolvedValueOnce({
+        data: { user: session.user, session },
+        error: null,
+      })
+      .mockResolvedValueOnce({
+        data: { user: null, session: null },
+        error: null,
+      });
+
+    render(
+      <AuthProvider>
+        <AuthProbe />
+      </AuthProvider>,
+    );
+
+    await waitFor(() => {
+      expect(readAuthState().loading).toBe(false);
+    });
+
+    act(() => {
+      authStateChangeHandler?.('SIGNED_IN');
+    });
+
+    await waitFor(() => {
+      expect(readAuthState().userId).toBe('practitioner-1');
+    });
+
+    act(() => {
+      authStateChangeHandler?.('TOKEN_REFRESHED');
+    });
+
+    await waitFor(() => {
+      expect(readAuthState().userId).toBeNull();
+    });
+
+    expect(signOutMock).not.toHaveBeenCalled();
+  });
+
   it('clears session when verify reports a refresh-token failure', async () => {
     const session = makeSession('practitioner-1');
