+++ apps/web/src/lib/auth/auth-callback-fragment-helpers.ts
@@ -16,20 +16,6 @@ export function parseImplicitHashParams(hash: string): Record<string, string> {
 }
 
 /**
- * True when `err` is a Supabase Auth API error from `getUser()` / `setSession()` (not a config throw).
- *
- * @param err - Value from `catch` or `getUser()` `error`.
- */
-export function isSupabaseAuthApiError(err: unknown): boolean {
-  return (
-    typeof err === 'object' &&
-    err !== null &&
-    '__isAuthError' in err &&
-    (err as { __isAuthError?: boolean }).__isAuthError === true
-  );
-}
-
-/**
  * True when `createBrowserClient()` failed because URL / publishable key env is missing or invalid
  * (throws from `apps/web` Supabase env helpers or `@abstrack/supabase` publishable key guards).
  *
